Move ingest business logic out of router
